simplify-ios-struct
